@@ -44,9 +50,10 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <sonar.organization>tacianosilva-github</sonar.organization>
         <sonar.language>java</sonar.language>
+        <sonar.java.source>java</sonar.java.source>
         <sonar.exclusions>**/*.properties,**/*.xml</sonar.exclusions>
         <!-- The main class to start by executing "java -jar" -->
         <start-class>br.ufrn.dct.apf.APFApplication</start-class>
@@ -258,40 +264,42 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-project-info-reports-plugin</artifactId>
-                <version>3.0.0</version>
+                <version>3.4.3</version>
                 <configuration>
                     <dependencyDetailsEnabled>true</dependencyDetailsEnabled>
                 </configuration>
-                <!-- <reportSets> <reportSet> <reports> <report>dependencies</report> 
+                <!-- <reportSets> <reportSet> <reports> <report>dependencies</report>
                     <report>scm</report> </reports> </reportSet> </reportSets> -->
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-javadoc-plugin</artifactId>
-                <version>3.0.0</version>
+                <version>3.5.0</version>
                 <configuration>
-                    <source>1.8</source>
+                    <source>17</source>
                 </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-report-plugin</artifactId>
-                <version>2.22.0</version>
+                <version>3.0.0</version>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-pmd-plugin</artifactId>
-                <version>3.10.0</version>
+                <version>3.20.0</version>
             </plugin>
+            <!--
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>findbugs-maven-plugin</artifactId>
                 <version>3.0.5</version>
             </plugin>
+            -->
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
-                <version>0.8.1</version>
+                <version>0.8.10</version>
             </plugin>
         </plugins>
     </reporting>
