@@ -2,12 +2,12 @@ package br.ufrn.dct.apf.model;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "attribution")
@@ -22,11 +22,11 @@ public class Attribution implements Serializable {
      * Regra para definir permissões membros de um projeto.
      */
     public static final String PROJECT_MANAGER = "PROJECT MANAGER";
-    
+
     public static final String PROJECT_MEMBER = "PROJECT MEMBER";
 
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "attrib_id")
     private int id;
 
