@@ -18,12 +18,12 @@ public abstract class AbstractController {
     private User overridenCurrentUser;
 
     @Autowired
-    private UserService userService;
+    protected UserService userService;
 
     protected void setUserAuth(ModelAndView modelAndView) {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         User user = userService.findUserByEmail(auth.getName());
-        modelAndView.addObject("userName", user.getName() + " " + user.getLastName() + " (" + user.getEmail() + ")");
+        modelAndView.addObject("userName", user.getFirstName() + " " + user.getLastName() + " (" + user.getEmail() + ")");
         modelAndView.addObject("authorities", getRoles(user));
     }
 
