@@ -8,6 +8,6 @@ import br.ufrn.dct.apf.model.User;
 @Repository("userRepository")
 public interface UserRepository extends JpaRepository<User, Long> {
 
-     User findByEmail(String email);
+     public User findByEmail(String email);
 
 }
